requirements.txt
Flask==3.0.0
Pyrebase4==4.6.0
python-dotenv==1.0.0